MA-10: refactoring
