+++ src/components/Dialogs/DialogItem/DialogItem.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import s from './../Dialogs.module.css'
+import { NavLink } from 'react-router-dom'
+
+const DialogItem = props => {
+	let path = '/dialogs/*' + props.id
+	return (
+		<div className={s.dialog}>
+			<NavLink to={path}>{props.name}</NavLink>
+		</div>
+	)
+}
+
+export default DialogItem
