+++ src/components/Dialogs/Dialogs.jsx
@@ -1,19 +1,8 @@
 import React from 'react'
 import s from './Dialogs.module.css'
-import { NavLink } from 'react-router-dom'
+import DialogItem from './DialogItem/DialogItem'
+import Message from './Message/Message'
 
-const DialogItem = props => {
-	let path = '/dialogs/*' + props.id
-	return (
-		<div className={s.dialog}>
-			<NavLink to={path}>{props.name}</NavLink>
-		</div>
-	)
-}
-
-const Message = props => {
-	return <div className={s.message}>{props.message}</div>
-}
 const Dialogs = props => {
 	let dialogs = [
 		{ id: 1, name: 'Valeria' },
